isspace() and tolower() are int->int as per the specs. These are generally implemented as table lookups and the compiler gives an error about a character being used as an table index

+++ trunk/gui/dir_dlg.c
@@ -325,7 +325,7 @@ CheckAccessAndChoose(AG_DirDlg *dd)
 	char *s;
 
 	for (s = &dd->cwd[0]; *s != '\0'; s++) {
-		if (!isspace(*s))
+		if (!isspace((int) *s))
 			break;
 	}
 	if (*s == '\0')
